pyproject.toml
[tool.poetry]
name = "metac-bot-template"
version = "0.1.0"
description = ""
authors = ["Vasile Popescu <elisescu@elisescu.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
requests = "^2.32.3"
asknews = "^0.9.1"
numpy = "^1.26.0"
openai = "^1.57.4"
python-dotenv = "^1.0.1"
forecasting-tools = "^0.2.23"
langchain-core = "^0.3.63"
jsonschema = "^4.24.0"
pytest = "^8.4.0"
bleach = "^6.1.0"
cryptography = "^42.0.0"
aioredis = "^2.0.1"
aiohttp = "^3.9.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest-cov = "^6.1.1"
ruff = "^0.11.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
